+++ aoc2019/aoc2019/param_mode.py
