+++ aoc2019/tests/test_day_05.py
@@ -1,5 +1,6 @@
-from aoc2019 import day_05
 import pytest
+
+from aoc2019.intcode import IntCode
 
 
 @pytest.mark.parametrize(
@@ -11,24 +12,24 @@ import pytest
     ]
 )
 def test_d2(input_file, user_input, expected_result):
-    intcode = day_05.IntCode()
+    intcode = IntCode()
     intcode.get_input(input_file)
     assert intcode.diagnostic_program(user_input) == expected_result
 
 def test_d5p1():
-    intcode = day_05.IntCode()
+    intcode = IntCode()
     intcode.get_input("tests/data/input_d5")
     intcode.diagnostic_program(1)
     assert intcode.result == 16489636
 
 def test_d5_examples():
-    intcode = day_05.IntCode()
+    intcode = IntCode()
     intcode.data = [3, 12, 6, 12, 15, 1, 13, 14, 13, 4, 13, 99, -1, 0, 1, 9]
     intcode.diagnostic_program(0)
     assert intcode.result == 0
 
 def test_d5p2():
-    intcode = day_05.IntCode()
+    intcode = IntCode()
     intcode.get_input("tests/data/input")
     intcode.diagnostic_program(5)
     assert intcode.result == 9386583
